requirements.txt
streamlit==1.28.1
openai==1.55.3
requests==2.31.0
pandas==2.1.3
pdfkit==1.0.0
xhtml2pdf==0.2.11
python-dotenv==1.0.0
markdown==3.5.1 